adding multichats panel
