+++ packages/jupyter-chat/src/index.ts
@@ -13,3 +13,4 @@ export * from './registers';
 export * from './selection-watcher';
 export * from './types';
 export * from './widgets';
+export * from './multiChatPanel';
